Allow partial support for `bevy_log` in `no_std`
