+++ crates/bevy_dylib/src/lib.rs
@@ -53,6 +53,8 @@
 //! use bevy_dylib;
 //! ```
 
+#![no_std]
+
 // Force linking of the main bevy crate
 #[expect(
     unused_imports,
