+++ crates/bevy_internal/Cargo.toml
@@ -10,7 +10,9 @@ keywords = ["game", "engine", "gamedev", "graphics", "bevy"]
 categories = ["game-engines", "graphics", "gui", "rendering"]
 
 [features]
+tracing = ["bevy_log/tracing"]
 trace = [
+  "tracing",
   "bevy_app/trace",
   "bevy_asset?/trace",
   "bevy_core_pipeline?/trace",
@@ -21,10 +23,18 @@ trace = [
   "bevy_render?/trace",
   "bevy_winit?/trace",
 ]
-trace_chrome = ["bevy_log/tracing-chrome"]
-trace_tracy = ["bevy_render?/tracing-tracy", "bevy_log/tracing-tracy"]
-trace_tracy_memory = ["bevy_log/trace_tracy_memory"]
-detailed_trace = ["bevy_ecs/detailed_trace", "bevy_render?/detailed_trace"]
+trace_chrome = ["tracing", "bevy_log/tracing-chrome"]
+trace_tracy = [
+  "tracing",
+  "bevy_render?/tracing-tracy",
+  "bevy_log/tracing-tracy",
+]
+trace_tracy_memory = ["tracing", "bevy_log/trace_tracy_memory"]
+detailed_trace = [
+  "tracing",
+  "bevy_ecs/detailed_trace",
+  "bevy_render?/detailed_trace",
+]
 
 sysinfo_plugin = ["bevy_diagnostic/sysinfo_plugin"]
 
@@ -293,6 +303,7 @@ std = [
   "bevy_ecs/std",
   "bevy_input/std",
   "bevy_input_focus?/std",
+  "bevy_log/std",
   "bevy_math/std",
   "bevy_platform_support/std",
   "bevy_reflect/std",
@@ -344,6 +355,7 @@ async_executor = [
 # Note this is currently only applicable on `wasm32` architectures.
 web = [
   "bevy_app/web",
+  "bevy_log/web",
   "bevy_platform_support/web",
   "bevy_reflect/web",
   "bevy_tasks/web",
@@ -385,8 +397,7 @@ bevy_utils = { path = "../bevy_utils", version = "0.16.0-dev", default-features 
 ] }
 bevy_tasks = { path = "../bevy_tasks", version = "0.16.0-dev", default-features = false }
 
-# bevy (std required)
-bevy_log = { path = "../bevy_log", version = "0.16.0-dev", optional = true }
+bevy_log = { path = "../bevy_log", version = "0.16.0-dev", default-features = false }
 
 # bevy (optional)
 bevy_a11y = { path = "../bevy_a11y", optional = true, version = "0.16.0-dev", features = [
