+++ crates/bevy_internal/src/prelude.rs
@@ -1,13 +1,9 @@
 #[doc(hidden)]
 pub use crate::{
-    app::prelude::*, ecs::prelude::*, input::prelude::*, math::prelude::*,
+    app::prelude::*, ecs::prelude::*, input::prelude::*, log::prelude::*, math::prelude::*,
     platform_support::prelude::*, reflect::prelude::*, time::prelude::*, transform::prelude::*,
     utils::prelude::*, DefaultPlugins, MinimalPlugins,
 };
-
-#[doc(hidden)]
-#[cfg(feature = "bevy_log")]
-pub use crate::log::prelude::*;
 
 #[doc(hidden)]
 #[cfg(feature = "bevy_window")]
