+++ crates/bevy_transform/Cargo.toml
@@ -70,6 +70,8 @@ std = [
   "alloc",
   "bevy_app?/std",
   "bevy_log",
+  "bevy_log?/std",
+  "bevy_log?/tracing",
   "bevy_ecs?/std",
   "bevy_math/std",
   "bevy_reflect?/std",
